Back to NZ1, unsplit e_room_exit
